+++ plugins/kboard/skin/thumbnail/list.php
@@ -63,6 +63,7 @@
 					<td class="kboard-list-thumbnail"></td>
 					<td class="kboard-list-title"><div class="cut_strings">
 							<a href="<?=$url->set('uid', $content->uid)->set('mod', 'document')->toString()?>"><?=$content->title?></a>
+							<?=$content->getCommentsCount();?>
 						</div></td>
 					<td class="kboard-list-user"><?=$content->member_display?></td>
 					<td class="kboard-list-date"><?=date("Y.m.d", strtotime($content->date))?></td>
@@ -75,11 +76,7 @@
 					<td class="kboard-list-thumbnail"><?php if($content->thumbnail_file):?><img src="<?=get_site_url() . $content->thumbnail_file?>" style="max-width: 100px;" alt="<?=$content->thumbnail_name?>"><?php else:?>이미지 없음<?php endif?></td>
 					<td class="kboard-list-title"><div class="cut_strings">
 							<a href="<?=$url->set('uid', $content->uid)->set('mod', 'document')->toString()?>"><?=$content->title?></a>
-							<?php 
-							$commentList = new CommentList($content->uid);
-							$commentsCount = $commentList->getCount();
-							if($commentsCount) echo "($commentsCount)";
-							?>
+							<?=$content->getCommentsCount();?>
 						</div></td>
 					<td class="kboard-list-user"><?=$content->member_display?></td>
 					<td class="kboard-list-date"><?=date("Y.m.d", strtotime($content->date))?></td>